+++ app/src/main/java/io/gnosis/safe/utils/AppUtils.kt
@@ -5,7 +5,6 @@ import android.content.Context
 import android.content.DialogInterface
 import android.content.Intent
 import android.content.res.Resources
-import android.graphics.Color
 import android.view.LayoutInflater
 import androidx.annotation.ColorRes
 import androidx.annotation.StringRes
@@ -13,8 +12,6 @@ import io.gnosis.safe.R
 import io.gnosis.safe.databinding.DialogRemoveBinding
 import io.gnosis.safe.qrscanner.QRCodeScanActivity
 import pm.gnosis.models.AddressBookEntry
-import pm.gnosis.svalinn.common.utils.getColorCompat
-import timber.log.Timber
 
 fun handleQrCodeActivityResult(
     requestCode: Int,
@@ -86,14 +83,5 @@ fun showConfirmDialog(
         cancelColor = R.color.primary,
         title = if (title == null) null else context.resources.getString(title)
     ).show()
-}
-
-fun String.toColor(context: Context, failsafeColor: Int = R.color.white): Int {
-    return try {
-        Color.parseColor(this)
-    } catch (e: Exception) {
-        Timber.e(e)
-        context.getColorCompat(failsafeColor)
-    }
 }
